[CSL-1196] Move most of light wallet stuff into Pos.Wallet.Light
