+++ src/Pos/Wallet.hs
@@ -1,13 +1,11 @@
 module Pos.Wallet
        ( module Pos.Wallet.KeyStorage
-       , module Pos.Wallet.Launcher
        , module Pos.Wallet.WalletMode
        , module Pos.Wallet.SscType
        , module Pos.Wallet.Web
        ) where
 
 import           Pos.Wallet.KeyStorage
-import           Pos.Wallet.Launcher
 import           Pos.Wallet.SscType
 import           Pos.Wallet.WalletMode
 import           Pos.Wallet.Web
