+++ src/Pos/Wallet/Light.hs
@@ -1,0 +1,3 @@
+-- | Light wallet.
+
+{-# OPTIONS_GHC -F -pgmF autoexporter #-}
