+++ src/Pos/Wallet/Light/Launcher.hs
@@ -1,0 +1,3 @@
+-- | Light wallet launcher.
+
+{-# OPTIONS_GHC -F -pgmF autoexporter #-}
