+++ src/Pos/Wallet/Light/State.hs
@@ -1,0 +1,7 @@
+module Pos.Wallet.Light.State
+       ( module Pos.Wallet.Light.State.Core
+       , module Pos.Wallet.Light.State.State
+       ) where
+
+import           Pos.Wallet.Light.State.Core
+import           Pos.Wallet.Light.State.State
