+++ src/Pos/Wallet/Light/Web.hs
@@ -1,2 +1,3 @@
--- Pos.Wallet.Launcher
+-- | Web part of light wallet
+
 {-# OPTIONS_GHC -F -pgmF autoexporter #-}
