--- src/Pos/Wallet/State.hs
@@ -1,7 +1,0 @@
-module Pos.Wallet.State
-       ( module Pos.Wallet.State.Core
-       , module Pos.Wallet.State.State
-       ) where
-
-import           Pos.Wallet.State.Core
-import           Pos.Wallet.State.State
